Tweaking og GUI
After a test flight, I found it difficult to find the right buttons when the AP button is aligned with the other buttons, so I moved it. I made the knob slightly bigger so that we can possibly fit longer texts. I also refined the King Air C90B configuration. Finally, I removed the legacy GUI, since the performance is so bad and I will not being using it.